packages/workforce/workforce/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workforce"
version = "0.1.0"
description = "Multi-tenant workforce management MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "a2wsgi>=1.10",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.27",
    "ollama>=0.2",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
        "pytest>=8.0",
"ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
